feat(*): subfield selection + better args + prep move queries
Subfield selection is now possible and for now still required. I still need to do some work on
making in optional through making some Fragments available. I've done a lot of refactoring in both
the data structure to make it easier to parse, and also some refactors that were required for
subfield selection. I just stumbled on how it worked so I went ahead and fully implemented it. Also
did some preps on move queries by creating the structure for it.

resolves #14