build(package,docker): better deployment setup
Moved most dependencies to dev so they are not installed with the typings and made Docker install
those into the environment as well. This ensures a minimal publishing typings bundle. Also told the
GraphQL Playground that it's on /api and not on /playground so it should default to that properly.