feat(*): implement requested field aware parsing
Previously when requested only a subsets of fields the API would still parse all data and let
GraphQL return the subset of requested data. This has been completely reworked and the API will now
always return only that subset of data that the user requested *before* it lets GraphQL figure this
out. This will provide performance improvement, in particular for the TypeMatchup and Learnset
queries where field-specific parsing is used to completely skip on processor intensive tasks if that
data is not being requested

BREAKING CHANGE: API has been majorly refactored and this is mostly non-breaking for client-side
typings, except for the fact that most resolver queries got a new parameter which will show up in
typings. So if you were specifically using the "Query" type then you will have some breaking
changes!