feat(*): rework parsing of aliases
Since its creation this library has supported a good set of aliases to match those used in the
Smogon community. However the way they had been implemented up until now they weren't ever really
being reliably used. This has now changed as the way it is determined whether a query matches an
alias is now much more streamlined

BREAKING CHANGE: While this should not be a breaking change to a lot of people, the types for
aliasses have drastically changed so this should be considered a breaking change.